Add model checkpoints for image NN
In this commit, I'm  adding the model checkpoints for the Neural Networks architecture trained on the Kuzushiji_MNIST dataset. The checkpoints were obtained after training the model for 500 epochs  a batch size of 128. We've used the Adam optimizer with default parameters and implemented a learning rate scheduler to improve convergence.

Summary of changes:
- Added checkpoint files for NN model
- Trained the model on Kuzushiji_MNIST dataset for 500 epochs
- batch size of 128
- Utilized Adam optimizer with default parameters
- Implemented learning rate scheduler for better convergence

Model Performance:
- Achieved an accuracy of 90% on the test set